requirements.txt
Fiona
shapely
psycopg2
requests
Pillow
progressbar2
numpy
pytest
pyyaml
